커뮤니티 글 수정 -> 이미지 추가 ok -> 이미지 삭제 기능 ok (ajax 처리 필요) 댓글 삭제시 ajax 구현 필요

+++ src/main/java/zerogreen/eco/service/community/BoardImageServiceImpl.java
@@ -21,6 +21,7 @@ public class BoardImageServiceImpl implements BoardImageService {
     private final BoardImageRepository boardImageRepository;
     private final CommunityBoardRepository boardRepository;
 
+    /* 이미지 Entity List -> Dto List */
     @Override
     public List<ImageFileDto> findByBoardId(Long boardId) {
         return boardImageRepository.findByBoard(boardId).stream().map(ImageFileDto::new).collect(Collectors.toList());
@@ -30,7 +31,6 @@ public class BoardImageServiceImpl implements BoardImageService {
     @Override
     public void modifyImage(Long boardId, String storeName, String originalName, String path) {
         CommunityBoard communityBoard = boardRepository.findById(boardId).orElseThrow();
-
         boardImageRepository.save(new BoardImage(originalName, storeName, path, communityBoard));
     }
 
@@ -41,8 +41,10 @@ public class BoardImageServiceImpl implements BoardImageService {
 
         try {
             if (file.exists()) {
-                file.delete();
-                boardImageRepository.deleteById(imageId);
+                boolean delete = file.delete();
+                if (delete) {
+                    boardImageRepository.deleteById(imageId);
+                }
             }
         } catch (Exception e) {
             e.printStackTrace();
